use PageHelper for listDataset paging query

+++ wedpr-components/dataset/src/main/java/com/webank/wedpr/components/dataset/message/ListDatasetResponse.java
@@ -8,7 +8,7 @@ import lombok.Data;
 @Data
 @Builder
 public class ListDatasetResponse {
-    int totalCount;
+    long totalCount;
     boolean isLast;
     List<Dataset> content;
 }
